+++ src/kcse/client_api.hpp
@@ -93,6 +93,7 @@ namespace kcd2o::kcse
 		std::uint32_t property_operation_generation{};
 		std::uint32_t property_operation_success{};
 		char property_operation_message[text_capacity]{};
+		std::uint32_t simulation_mode{};
 	};
 
 	struct remote_player_view